android: use device tree component type along with source
The original intent of using device tree source as DT
component metadata was to help the firmware distinguish
between device trees and overlays.

While source is a useful piece of metadata, it is not very
suitable for detecting component types alone, since the same
source may provide different component types:

1. The DTBO partition may contain both VB_DTBOs and HLOS DTBOs
   to provide AVF device assignment overlay.
2. A FIT image can carry both DTBs and DTBOs.

This change extends metadata with explicit device tree type
to make sure we can support such cases in a future.

Also add PVM_VA_OVERLAY type to the protocol definition to be
supported by GBL in the future.

Bug: 353272981
Bug: 385690995
Test: booted sun after updating FW implementation
Change-Id: Ib1f2c869f07da2c0085540b90fe79228784b0e1b
Signed-off-by: Dmitrii Merkurev <dimorinny@google.com>