Handle bootimage with GKI certificate
GKI kernel image may be appended with an additional 16K GKI certificate.
The certificate is also included in the avb verification but not
reflected in the bootimg header. As a quick workaround, preloads an
additional 16k of data while loading boot_a/boot_b.

Also fixes an issue where android_load_verify_fixup is not checking
verification status of each preloaded partition.

Bug: 416000842

Change-Id: I5f24a6a4f0c2f7c023fc34a502fcde8f88fbe1db