Switch to letting libavb load partitions
It was observed that boot image may be appended with an addition GKI
certificate that gets included in libavb verification but not reflected
by boot image header. This reveals an issue that if we expect boot
images to have arbitrary payload appended. The current preload approach
will not be able to preload the correct amount of data for avb
verification. To avoid this potential problem, switch to letting libavb
to allocate and load partitions according to the size information from
hash descriptors, which is always correct.

This will also be necessary if we want to acheive parallel image
loading/hashing because the salt, hash algorithm information etc is
only available within libavb and thus most likely not feasible for any
preloading approach.

Bug: 416000842

Change-Id: I3a0cc26bde3018bfcc6a52490123eb2bd9e48fef