gbl: add libefi_types protocol support
Adds the framework for protocol support for both clients and providers.

The idea here is that we have a single Rust trait to represent a
protocol, and structs can implement that trait either as a client (e.g.
GBL) or a provider (e.g. a Rust UEFI firmware).

The provider side is optional, and we expect that most vendors will
continue to use C for the time being, but for those that do choose Rust
this will provide better consistency with GBL, as well as hiding the
vast majority of unsafe glue between the C and Rust ABIs.

For GBL, this won't affect the logic too much, it will mostly be a
moving of the unsafe Rust/C glue out into this more focused library.

Bug: b/413021069
Change-Id: Id11224de60bc4822f26176923bc94b983b61e7b3