gbl: switch storage APIs to use UninitSlice
There was a safety error in our MaybeUninit APIs which allowed
casting &mut [u8] to &mut [MaybeUninit<u8>] in safe code. This is
unsound because then writing MaybeUninit::uninit() to any of the
bytes can trigger UB since now the &[u8] has been "uninitialized"
and is no longer valid.

It's very unlikely we were actually triggering this case, but it does
show some of the subtleties of MaybeUninit. This CL instead switches to
the bytes crate UninitSlice, which basically does the things we were
trying to do anyway but is designed to explicitly not allow this exact
error of uninitializing a &[u8].

This is roughly net neutral in terms of code complexity, but eliminates
the safety error with minimal additional `unsafe` code blocks, which is
worthwhile.

Bug: b/421469920
Change-Id: Ib9c8c5d397d089aa3d607ba7bfecfa0c4622cec3