gbl: use interior mutability for protocols
UEFI protocols may need to be opened by multiple clients at the same
time, which the previous &mut implementation did not allow. This CL
switches to interior mutability, so that multiple clients can open a
protocol but the protocol implementation has to ensure safe access.

Safe access can be a little tricky in the UEFI environment, so this CL
also adds documentation indicating how to safely control access and
avoid race conditions.

Bug: b/413021069
Change-Id: Ifd4c226d5e0e1cf4da5986b10483668a3dda7a53