Log missing protocols and NULL protocol methods
It is helpful for developers, both those using and modifying GBL, to
be alerted when required protocols are not present in the UEFI
environment and when protocols are present but methods are not yet
implemented (i.e. the function pointer is NULL).

Within GBL, all protocols are now tagged as either 'Mandatory' or
'Optional'. If GBL attempts to open a 'Mandatory' protocol that is not
present, the error is logged before taking any other action.
The absence of 'Optional' protocols is also logged for 'dev' builds.

When attempting an EFI call, NULL method pointers are logged.

Tests: new unit tests pass,
       missing protocols and methods are logged in QEMU.

Bugs: b/432104967 b/431865972

Change-Id: I1540f2389f106b74d3d8418d7005d99534dd043a