tests: fix test data path for @gbl.
Now that gbl is a bazel_dep(), its canonical repository name becomes
gbl+. Update test data paths to reflect it.

The more robust way would be to get it from Bazel's context. However,
I do not have the bandwidth to properly fix this, so just providing
a band-aid solution for now.

Bug: 413506174
Change-Id: I490edbaf57f86de0d886061f857d32b22695abca