Support partial receive of fastboot packet
GBL fastboot transport receives data in the unit of packet. For media
such as TCP, a packet can be the entire download data. This makes it
difficult to monitor or report download progress which has been useful
for diagnosis. The CL extends the transport interface to allow partial
receive of fastboot packet.

Bug: 456179055
Change-Id: Ic0a649b9f3ae1185728692dd3f29e18e66e0170f