Add support for "fastboot flashing get_unlock_ability"
Augment the return flags available to
`GblAvbProtocol.read_device_status()` to indicate whether the device
can be unlocked.

Add the "get_unlock_ability" subcommand to fastboot flashing in the
fastboot engine and route it to check for the presence of the flag.

Remove the GblFastbootProtocol `get_lock()` method.
Move the GblFastbootProtocol `set_lock()` method to GblAvbProtocol and
rename it to `set_lock_state()`.

Bug: b/462540091

Change-Id: I65ac7eec511df68e02b9646c13630c1166480e11