GBL: Fix permission error in gblsigntool_test
There was a permission error when running the test case because the test
uses `shutils.copy` to generate test data dynamically. The issue is that
the source file mode is also copied to the destination file, and the
made the destination file RO, thus opening the destination file in write
mode would fail.
Replace the `shutils.copy` with plain file creation operations.

Test: ./tools/bazel test @gbl//tests
Change-Id: I031a2a63e19ec0fc867daa0984a472d3e9c34d7f